+++ modules/server/src/main/java/org/apache/airavata/server/ServerMain.java
@@ -60,7 +60,7 @@ public class ServerMain {
 	private static final String API_SERVER = "apiserver";
 	private static final String CREDENTIAL_STORE = "credentialstore";
 	private static final String REGISTRY_SERVER = "regserver";
-	private static final String SHARING_SERVER = "sharing_server";
+	//private static final String SHARING_SERVER = "sharing_server";
 	private static final String GFAC_SERVER = "gfac";
 	private static final String ORCHESTRATOR = "orchestrator";
 	private static final String USER_PROFILE = "user_profile";
@@ -115,7 +115,7 @@ public class ServerMain {
 			serverList.clear();
 			serverList.add(REGISTRY_SERVER);  // registry server should start before everything
 			serverList.add(CREDENTIAL_STORE); // credential store should start before api server
-			serverList.add(SHARING_SERVER);
+			//serverList.add(SHARING_SERVER);
 			serverList.add(API_SERVER);
 			serverList.add(ORCHESTRATOR);
 			serverList.add(GFAC_SERVER);
@@ -124,7 +124,7 @@ public class ServerMain {
 			serverList.clear();
             serverList.add(REGISTRY_SERVER);  // registry server should start before everything
             serverList.add(CREDENTIAL_STORE); // credential store should start before api server
-			serverList.add(SHARING_SERVER);
+			//serverList.add(SHARING_SERVER);
 			serverList.add(API_SERVER);
 			serverList.add(ORCHESTRATOR);
 		} else if (serverList.indexOf(EXECUTION) > -1) {
