+++ modules/user-profile/user-profile-core/src/main/java/org/apache/airavata/user/registry/core/utils/Committer.java
@@ -17,16 +17,11 @@
  * specific language governing permissions and limitations
  * under the License.
  *
- */
+*/
+package org.apache.airavata.user.registry.core.utils;
 
-package org.apache.airavata.registry.core.utils;
-public class JPAConstants {
-	public static final String KEY_JDBC_URL = "user.profile.catalog.registry.jdbc.url";
-	public static final String KEY_JDBC_USER = "user.profile.catalog.registry.jdbc.user";
-	public static final String KEY_JDBC_PASSWORD = "user.profile.catalog.registry.jdbc.password";
-	public static final String KEY_JDBC_DRIVER = "user.profile.catalog.registry.jdbc.driver";
-	public static final String KEY_DERBY_START_ENABLE = "user.profile.catalog.start.derby.server.mode";
-	public static final String VALIDATION_QUERY = "user.profile.catalog.validationQuery";
-	public static final String JPA_CACHE_SIZE = "user.profile.catalog.jpa.cache.size";
-	public static final String ENABLE_CACHING = "user.profile.catalog.cache.enable";
+@FunctionalInterface
+public interface Committer<T, R>  {
+
+    R commit(T t);
 }
