+++ modules/user-profile/user-profile-core/src/main/java/org/apache/airavata/user/registry/core/utils/JPAUtils.java
@@ -18,7 +18,7 @@
  * under the License.
  *
 */
-package org.apache.airavata.registry.core.utils;
+package org.apache.airavata.user.registry.core.utils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
